+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-18T05:53:50.141Z",
+    "2026-02-18T08:50:37.645Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {